refactor: move page components into dedicated feature modules
Move MeshGradientLab and BookmarkManagement pages into the features
directory to improve modularity and separation of concerns. Update
entrypoints to import from the new feature locations.
